completely redid the lift control task, made op control cleaner with tasks

+++ dr4b_passive_active_better/include/misc.h
@@ -1,21 +1,21 @@
 #ifndef MISC_H
 #define MISC_H
 
-int sgn(float x); //inputs a double and outputs its sign (+ or -) as an int
+int sgn(double x); //inputs a double and outputs its sign (+ or -) as an int
 struct PID { //a holder for multiple values used in PID
-    float error, integral, derivative, prevError, kP, kI, kD; //PID variables and constants
+    double error, integral, derivative, prevError, kP, kI, kD; //PID variables and constants
     bool useP, useI, useD; //0 to disable certain terms, 1 to enable
 };
-PID initPID(bool useP, bool useI, bool useD, float kP, float kI, float kD); //gives a created PID struct its values
-float runPID(PID *pid); //inputs a PID struct, does the actual PID math, then outputs a double
+PID initPID(bool useP, bool useI, bool useD, double kP, double kI, double kD); //gives a created PID struct its values
+double runPID(PID *pid); //inputs a PID struct, does the actual PID math, then outputs a double
 
-float joyValRemap(float joyVal); //optimizes the slow and fast motor values during driver control
+double joyValRemap(double joyVal); //optimizes the slow and fast motor values during driver control
 
 //buttons and joysticks
-float lY();
-float lX();
-float rY();
-float rX();
+double lY();
+double lX();
+double rY();
+double rX();
 bool l1();
 bool l2();
 bool r1();
